payments/payments/requirements.txt
asgiref==3.8.1
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
Django==5.0.3
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.1
googleapis-common-protos==1.69.2
grpcio==1.71.0
idna==3.10
importlib-metadata==6.11.0
opentelemetry-api==1.22.0
opentelemetry-distro==0.52b1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-django==0.43b0
opentelemetry-instrumentation-logging==0.43b0
opentelemetry-instrumentation-requests==0.52b1
opentelemetry-instrumentation-wsgi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
packaging==24.2
protobuf==4.25.7
requests==2.32.3
setuptools==80.3.0
six==1.17.0
sqlparse==0.5.3
stripe==11.0.0
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.3.0
wrapt==1.17.2
zipp==3.21.0
